Make AudioPlayer final
